Rename Render -> MazeRenderer, make it a proper class
As a proper class, the cairo surface and context are held over the lifetime
of the MazeRenderer.  Not so many variables are passed between methods now.

These steps will make it easier to use MazeRenderer to redraw a portion of
the maze.

Pulled Arguments.h out of the library -- they were only used by Renderer
and they really didn't need to be.