Refactoring to better control maze generation
Create 'MazeApp' class that ties together a MazeModel, MazeMaker, and
MazeRenderer.  Call MazeApp::iterate to progress one step towards
completion of the creation of the maze.

Expose getting the maze image surface (bitmap) from MazeRenderer so it
can be drawn to the screen.

Also, if there are empty cells in the maze, draw them with a gray
background.